Add type in place of any and add default-element in container to config
