+++ builder/src/components/custom-components/container.tsx
@@ -6,6 +6,7 @@ import RenderItem from "utils/render-item";
 import IItems from "interfaces/items";
 import IColor from "interfaces/color";
 import "styles/components.css";
+import defaultItem from "config/default-element-container";
 
 interface IContainer {
   item: IItems;
@@ -60,27 +61,7 @@ const Container: FC<IContainer> = ({
   elementConfig,
   setElementConfig,
 }) => {
-  // default grid layout element for empty container
-  const defaultItem = {
-    i: "Dop",
-    link: "",
-    minW: 1,
-    name: "Text",
-    style: {
-      color: { r: "0", g: "0", b: "0", a: "100" },
-      backgroundColor: { r: "0", g: "0", b: "0", a: "" },
-      fontWeight: "normal",
-      fontStyle: "normal",
-      textDecoration: "none",
-      justifyContent: "center",
-      fontSize: 16,
-    },
-    value: "Hover and click on drag to add components in container",
-    w: 12,
-    x: 0,
-    y: 0,
-    h: 2,
-  }
+  
   // to persist layout changes
   const onLayoutChange = (layout: Layout[]) => {
     let newItemsArr = layout.map((obj: IItems) => {
