+++ builder/src/config/default-element-container.ts
@@ -1,0 +1,24 @@
+// default grid layout element for empty container 
+// user in custom-components/container
+const defaultItem = {
+  i: "DefaultElement",
+  link: "",
+  minW: 1,
+  name: "Text",
+  style: {
+    color: { r: "0", g: "0", b: "0", a: "100" },
+    backgroundColor: { r: "0", g: "0", b: "0", a: "" },
+    fontWeight: "normal",
+    fontStyle: "normal",
+    textDecoration: "none",
+    justifyContent: "center",
+    fontSize: 16,
+  },
+  value: "Hover and click on drag to add components in container",
+  w: 12,
+  x: 0,
+  y: 0,
+  h: 2,
+}
+
+export default defaultItem
